requirements.txt
imageio==2.9.0
matplotlib==3.3.3
networkx==2.5
numpy==1.19.5
PyYAML==5.4.1
requests-oauthlib==1.3.0
torch==1.7.1
torch-cluster==1.5.9
torch-geometric==1.6.3
torch-scatter==2.0.5
torch-sparse==0.6.8
torch-spline-conv==1.2.1
torchvision==0.8.2
tqdm==4.58.0

